code value

+++ easy-trade/easy-tms/src/main/java/com/rongyi/tms/service/impl/dubbo/CommissionServiceImpl.java
@@ -86,13 +86,13 @@ public class CommissionServiceImpl implements CommissionService {
             if (params.containsKey("status")) {
                 List<Integer> statusList = new ArrayList<>();
 
-                if (params.get("status") == Constants.DrawApplyStatus.SEND) {
+                if ((int)params.get("status") == Constants.DrawApplyStatus.SEND) {
                     statusList.add(ConstantEnum.COMMISSION_STATUS_6.getCodeInt());
-                } else if (params.get("status") == Constants.DrawApplyStatus.PROCESSING) {
+                } else if ((int)params.get("status") == Constants.DrawApplyStatus.PROCESSING) {
                     statusList.add(ConstantEnum.COMMISSION_STATUS_1.getCodeInt());
                     statusList.add(ConstantEnum.COMMISSION_STATUS_2.getCodeInt());
                     statusList.add(ConstantEnum.COMMISSION_STATUS_3.getCodeInt());
-                } else if (params.get("status") == Constants.DrawApplyStatus.FAIL) {
+                } else if ((int)params.get("status") == Constants.DrawApplyStatus.FAIL) {
                     statusList.add(ConstantEnum.COMMISSION_STATUS_1_UNCHECK.getCodeInt());
                     statusList.add(ConstantEnum.COMMISSION_STATUS_2_UNCHECK.getCodeInt());
                 }
